Move shared macos code to talpid-macos
