+++ Cargo.toml
@@ -29,6 +29,7 @@ members = [
     "talpid-core",
     "talpid-dbus",
     "talpid-future",
+    "talpid-macos",
     "talpid-net",
     "talpid-openvpn",
     "talpid-openvpn-plugin",
